requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
jinja2==3.1.6
python-multipart==0.0.20
httpx[http2]==0.28.1
apscheduler==3.10.4
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==44.0.2
pydantic-settings==2.8.1
itsdangerous==2.2.0
pyjwt>=2.8.0
pyjwt
redis
pymysql
celery
openpyxl==3.1.5
qiniu>=7.14.0
Pillow>=10.0.0
pymupdf>=1.25.0
qdrant-client==1.17.1
